refactor: reorganize ui files
