+++ lib/ui/features/info/info_state.dart
@@ -1,0 +1,9 @@
+
+import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:tv_randshow/core/app/domain/interfaces/i_app_service.dart';
+import 'package:tv_randshow/core/app/ioc/locator.dart';
+
+final versionAppProvider = FutureProvider.autoDispose<String>((ref) async {
+  final IAppService appService = locator<IAppService>();
+  return appService.getVersion();
+});
