+++ lib/ui/features/info/views/privacy_policy_view.dart
