+++ lib/ui/router.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
-import 'package:tv_randshow/ui/views/loading_view.dart';
-import 'package:tv_randshow/ui/views/migration_view.dart';
-import 'package:tv_randshow/ui/views/privacy_policy_view.dart';
-import 'package:tv_randshow/ui/views/result_view.dart';
-import 'package:tv_randshow/ui/views/splash_view.dart';
-import 'package:tv_randshow/ui/views/tab_view.dart';
+import 'package:tv_randshow/ui/features/info/views/privacy_policy_view.dart';
+import 'package:tv_randshow/ui/features/init/splash_view.dart';
+import 'package:tv_randshow/ui/features/init/tab_view.dart';
+import 'package:tv_randshow/ui/features/migration/migration_view.dart';
+import 'package:tv_randshow/ui/features/random/loading_view.dart';
+import 'package:tv_randshow/ui/features/random/result_view.dart';
 
 class RoutePaths {
   static const String migraiton = 'migration';
