+++ lib/ui/widgets/error_icon.dart
